requirements.txt
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
deep-translator==1.11.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.1
outcome==1.3.0.post0
pillow==11.3.0
proglog==0.1.12
psutil==7.0.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.1
python-telegram-bot==22.2
PyWavelets==1.8.0
requests==2.32.4
rsa==4.9.1
scipy==1.16.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
Telethon==1.40.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
undetected-chromedrive==3.5.5